src/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
imbalanced-learn==0.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
multidict==6.0.4
nltk==3.8.1
numpy==1.24.2
pandas==1.5.3
pybind11==2.10.1
pyee==9.0.4
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
slackclient==2.9.4
slackeventsapi==3.0.1
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.5.0
Werkzeug==2.2.3
yarl==1.8.2
